+++ choco-solver/src/main/java/org/chocosolver/solver/variables/view/BoolEqView.java
@@ -34,6 +34,7 @@ import org.chocosolver.solver.Solver;
 import org.chocosolver.solver.exception.ContradictionException;
 import org.chocosolver.solver.exception.SolverException;
 import org.chocosolver.solver.variables.BoolVar;
+import org.chocosolver.solver.variables.VF;
 import org.chocosolver.solver.variables.Variable;
 import org.chocosolver.util.ESat;
 
@@ -70,7 +71,7 @@ public final class BoolEqView extends EqView implements BoolVar {
 
     @Override
     public BoolVar duplicate() {
-        return new BoolEqView(this.var);
+        return VF.eq(this.var);
     }
 
     @Override